Move `PrettyFormatOption` to break circular dependencies.
