+++ metricflow-semantics/tests_metricflow_semantics/mf_logging/test_pretty_print.py
@@ -10,11 +10,11 @@ from dbt_semantic_interfaces.implementations.elements.dimension import PydanticD
 from dbt_semantic_interfaces.type_enums import DimensionType
 from metricflow_semantics.experimental.ordered_set import FrozenOrderedSet
 from metricflow_semantics.helpers.string_helpers import mf_dedent, mf_indent
+from metricflow_semantics.mf_logging.format_option import PrettyFormatOption
 from metricflow_semantics.mf_logging.pretty_formattable import MetricFlowPrettyFormattable
 from metricflow_semantics.mf_logging.pretty_formatter import (
     MetricFlowPrettyFormatter,
     PrettyFormatContext,
-    PrettyFormatOption,
 )
 from metricflow_semantics.mf_logging.pretty_print import PrettyFormatDictOption, mf_pformat
 from metricflow_semantics.test_helpers.metric_time_dimension import MTD_SPEC_DAY
